taffy::compute::grid::alignment

Function align_tracks

pub(super) fn align_tracks(
    grid_container_content_box_size: f32,
    padding: Line<f32>,
    border: Line<f32>,
    tracks: &mut [GridTrack],
    track_alignment_style: AlignContent,
)

Align the grid tracks within the grid according to the align-content (rows) or justify-content (columns) property. This only does anything if the size of the grid is not equal to the size of the grid container in the axis being aligned.