taffy::compute::grid

Module explicit_grid

Helper functions for initialising GridTrack’s from styles This mainly consists of evaluating GridAutoTracks

Enums

AutoRepeatStrategy 🔒

The auto-repeat fit strategy to use

Functions

compute_explicit_grid_size_in_axis 🔒

Compute the number of rows and columns in the explicit grid

create_implicit_tracks 🔒

Utility function for repeating logic of creating implicit tracks

initialize_grid_tracks 🔒

Resolve the track sizing functions of explicit tracks, automatically created tracks, and gutters given a set of track counts and all of the relevant styles