taffy::compute::grid

Module implicit_grid

This module is not required for spec compliance, but is used as a performance optimisation to reduce the number of allocations required when creating a grid.

Functions

child_min_line_max_line_span 🔒

Helper function for compute_grid_size_estimate Produces a conservative estimate of the greatest and smallest grid lines used by a single grid item

compute_grid_size_estimate 🔒

Estimate the number of rows and columns in the grid This is used as a performance optimisation to pre-size vectors and reduce allocations. It also forms a necessary step in the auto-placement

get_known_child_positions 🔒

Iterate over children, producing an estimate of the min and max grid lines along with the span of each item