taffy::compute::grid::implicit_grid

Function child_min_line_max_line_span

fn child_min_line_max_line_span<S: CheapCloneStr>(
    line: Line<GridPlacement<S>>,
    explicit_track_count: u16,
) -> (OriginZeroLine, OriginZeroLine, u16)

Helper function for compute_grid_size_estimate Produces a conservative estimate of the greatest and smallest grid lines used by a single grid item

Values are returned in origin-zero coordinates