taffy::compute::grid::track_sizing

Struct ItemBatcher

struct ItemBatcher {
    axis: AbstractAxis,
    index_offset: usize,
    current_span: u16,
    current_is_flex: bool,
}

Takes an axis, and a list of grid items sorted firstly by whether they cross a flex track in the specified axis (items that don’t cross a flex track first) and then by the number of tracks they cross in specified axis (ascending order).

Fields
axis: AbstractAxis

The axis in which the ItemBatcher is operating. Used when querying properties from items.

index_offset: usize

The starting index of the current batch

current_span: u16

The span of the items in the current batch

current_is_flex: bool

Whether the current batch of items cross a flexible track

Implementations

impl ItemBatcher

fn new(axis: AbstractAxis) -> Self

Create a new ItemBatcher for the specified axis

fn next<'items>( &mut self, items: &'items mut [GridItem], ) -> Option<(&'items mut [GridItem], bool)>

This is basically a manual version of Iterator::next which passes items in as a parameter on each iteration to work around borrow checker rules