taffy::compute::grid::types

Module coordinates

Taffy uses two coordinate systems to refer to grid lines (the gaps/gutters between rows/columns):

Structs

GridLine

Represents a grid line position in “CSS Grid Line” coordinates

OriginZeroLine

Represents a grid line position in “OriginZero” coordinates

Constants

MAX_GRID_TRACKS 🔒

The maximum number of tracks in each direction from the start of the explicit grid (line 0 in OriginZero coordinates), and the maximum span of a single grid item. This limits the grid to 2 * MAX_GRID_TRACKS total tracks in each axis (including explicit tracks, which count against the positive limit). Grids larger than this limit are clamped.

MAX_OZ_LINE 🔒

The highest valid grid line in OriginZero coordinates

MIN_OZ_LINE 🔒

The lowest valid grid line in OriginZero coordinates

Traits

GridCoordinate

A trait for the different coordinates used to define grid lines.