Struct taffy::geometry::Point

pub struct Point<T> {
    pub x: T,
    pub y: T,
}

A 2-dimensional coordinate.

When used in association with a Rect, represents the top-left corner.

Fields
x: T

The x-coordinate

y: T

The y-coordinate

Implementations

impl Point<f32>

pub const ZERO: Self = _

A Point with values (0,0), representing the origin

impl Point<Option<f32>>

pub const NONE: Self = _

A Point with values (None, None)

impl<T> Point<T>

pub fn map<R, F>(self, f: F) -> Point<R>
where F: Fn(T) -> R,

Applies the function f to both the x and y

This is used to transform a Point<T> into a Point<R>.

pub fn get(self, axis: AbstractAxis) -> T

Gets the extent of the specified layout axis Whether this is the width or height depends on the GridAxis provided

pub fn transpose(self) -> Point<T>

Swap x and y components

pub fn set(&mut self, axis: AbstractAxis, value: T)

Sets the extent of the specified layout axis Whether this is the width or height depends on the GridAxis provided

impl<T: TaffyZero> Point<T>

pub const fn zero() -> Self

Returns a Point where both the x and y values are the zero value of the contained type (e.g. 0.0, Some(0.0), or Dimension::Length(0.0))

impl<T: TaffyAuto> Point<T>

pub const fn auto() -> Self

Returns a Point where both the x and y values are the auto value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T: TaffyMinContent> Point<T>

pub const fn min_content() -> Self

Returns a Point where both the x and y values are the min_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T: TaffyMaxContent> Point<T>

pub const fn max_content() -> Self

Returns a Point where both the x and y values are the max_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T: TaffyFitContent> Point<T>

pub fn fit_content(argument: LengthPercentage) -> Self

Returns a Point with x and y set to the same fit-content(…) value with the given argument.

impl<T: FromLength> Point<T>

pub fn length<Input: Into<f32> + Copy>(value: Input) -> Self

Returns a Point where x and y values are the same given absolute length

impl<T: FromPercent> Point<T>

pub fn percent<Input: Into<f32> + Copy>(percent_value: Input) -> Self

Returns a Point where both the x and y values are the constant percent value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Length(2.1))