Struct taffy::geometry::Rect

pub struct Rect<T> {
    pub left: T,
    pub right: T,
    pub top: T,
    pub bottom: T,
}

An axis-aligned UI rectangle

Fields
left: T

This can represent either the x-coordinate of the starting edge, or the amount of padding on the starting side.

The starting edge is the left edge when working with LTR text, and the right edge when working with RTL text.

right: T

This can represent either the x-coordinate of the ending edge, or the amount of padding on the ending side.

The ending edge is the right edge when working with LTR text, and the left edge when working with RTL text.

top: T

This can represent either the y-coordinate of the top edge, or the amount of padding on the top side.

bottom: T

This can represent either the y-coordinate of the bottom edge, or the amount of padding on the bottom side.

Implementations

impl<T: Add> Rect<T>

pub fn grid_axis_sum(self, axis: AbsoluteAxis) -> <T as Add>::Output

Get either the width or height depending on the AbsoluteAxis passed in

impl<T> Rect<T>

pub fn map<R, F>(self, f: F) -> Rect<R>where
F: Fn(T) -> R,

Applies the function f to the left, right, top, and bottom properties

This is used to transform a Rect<T> into a Rect<R>.

pub fn horizontal_components(self) -> Line<T>

Returns a Line<T> representing the left and right properties of the Rect

pub fn vertical_components(self) -> Line<T>

Returns a Line<T> containing the top and bottom properties of the Rect

impl Rect<f32>

pub const ZERO: Rect<f32> = _

Creates a new Rect with 0.0 as all parameters

pub const fn new(start: f32, end: f32, top: f32, bottom: f32) -> Self

Creates a new Rect

impl Rect<Dimension>

pub const fn from_points(start: f32, end: f32, top: f32, bottom: f32) -> Self

Create a new Rect with Dimension::Points

pub const fn from_percent(start: f32, end: f32, top: f32, bottom: f32) -> Self

Create a new Rect with Dimension::Percent

impl<T: TaffyZero> Rect<T>

pub const fn zero() -> Self

Returns a Size where the left, right, top, and bottom values are all the zero value of the contained type (e.g. 0.0, Some(0.0), or Dimension::Points(0.0))

impl<T: TaffyAuto> Rect<T>

pub const fn auto() -> Self

Returns a Size where the left, right, top, and bottom values are all the auto value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T: TaffyMinContent> Rect<T>

pub const fn min_content() -> Self

Returns a Size where the left, right, top, and bottom values are all the min_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T: TaffyMaxContent> Rect<T>

pub const fn max_content() -> Self

Returns a Size where the left, right, top, and bottom values are all the max_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T: TaffyFitContent> Rect<T>

pub fn fit_content(argument: LengthPercentage) -> Self

Returns a Rect where the left, right, top and bottom values are all constant fit_content value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Points(2.1))

impl<T: FromPoints> Rect<T>

pub fn points<Input: Into<f32> + Copy>(points_value: Input) -> Self

Returns a Rect where the left, right, top and bottom values are all constant points value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Points(2.1))

impl<T: FromPercent> Rect<T>

pub fn percent<Input: Into<f32> + Copy>(percent_value: Input) -> Self

Returns a Rect where the left, right, top and bottom values are all constant percent value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Points(2.1))