Struct taffy::node::Taffy

pub struct Taffy { /* private fields */ }

A forest of UI Nodes, suitable for UI layout

Implementations

impl Taffy

pub fn new() -> Self

Creates a new Taffy

The default capacity of a Taffy is 16 nodes.

pub fn with_capacity(capacity: usize) -> Self

Creates a new Taffy that can store capacity nodes before reallocation

pub fn new_leaf(
&mut self,
style: Style,
measure: MeasureFunc
) -> Result<Node, Error>

Adds a new leaf node, which does not have any children

pub fn new_node(
&mut self,
style: Style,
children: &[Node]
) -> Result<Node, Error>

Adds a new node, which may have any number of children

pub fn clear(&mut self)

Removes all nodes

All associated [Id] will be rendered invalid.

pub fn remove(&mut self, node: Node)

Remove a specific Node from the tree

Its [Id] is marked as invalid.

pub fn set_measure(
&mut self,
node: Node,
measure: Option<MeasureFunc>
) -> Result<(), Error>

Sets the MeasureFunc of the associated node

pub fn add_child(&mut self, parent: Node, child: Node) -> Result<(), Error>

Adds a child Node under the supplied parent

pub fn set_children(
&mut self,
parent: Node,
children: &[Node]
) -> Result<(), Error>

Directly sets the children of the supplied parent

pub fn remove_child(&mut self, parent: Node, child: Node) -> Result<Node, Error>

Removes the child of the parent node

The child is not removed from the forest entirely, it is simply no longer attached to its previous parent.

pub fn remove_child_at_index(
&mut self,
parent: Node,
child_index: usize
) -> Result<Node, Error>

Removes the child at the given index from the parent

The child is not removed from the forest entirely, it is simply no longer attached to its previous parent.

pub fn replace_child_at_index(
&mut self,
parent: Node,
child_index: usize,
new_child: Node
) -> Result<Node, Error>

Replaces the child at the given child_index from the parent node with the new child node

The child is not removed from the forest entirely, it is simply no longer attached to its previous parent.

pub fn child_at_index(
&self,
parent: Node,
child_index: usize
) -> Result<Node, Error>

Returns the child Node of the parent node at the provided child_index

pub fn child_count(&self, parent: Node) -> Result<usize, Error>

Returns the number of children of the parent Node

pub fn children(&self, parent: Node) -> Result<Vec<Node>, Error>

Returns a list of children that belong to the [Parent]

pub fn set_style(&mut self, node: Node, style: Style) -> Result<(), Error>

Sets the Style of the provided node

pub fn style(&self, node: Node) -> Result<&Style, Error>

Gets the Style of the provided node

pub fn layout(&self, node: Node) -> Result<&Layout, Error>

Return this node layout relative to its parent

pub fn mark_dirty(&mut self, node: Node) -> Result<(), Error>

Marks the layout computation of this node and its children as outdated

pub fn dirty(&self, node: Node) -> Result<bool, Error>

Indicates whether the layout of this node (and its children) need to be recomputed

pub fn compute_layout(
&mut self,
node: Node,
size: Size<Number>
) -> Result<(), Error>

Updates the stored layout of the provided node and its children