Module taffy::prelude

Commonly used types

Re-exports

pub use crate::geometry::Line;

pub use crate::geometry::Rect;

pub use crate::geometry::Size;

pub use crate::layout::Layout;

pub use crate::node::Node;

pub use crate::node::Taffy;

pub use crate::style::Display;

pub use crate::style::Position;

pub use crate::style::Style;

pub use crate::style_helpers::auto;

pub use crate::style_helpers::fit_content;

pub use crate::style_helpers::max_content;

pub use crate::style_helpers::min_content;

pub use crate::style_helpers::percent;

pub use crate::style_helpers::points;

pub use crate::style_helpers::zero;

pub use crate::style_helpers::FromFlex;

pub use crate::style_helpers::FromPercent;

pub use crate::style_helpers::FromPoints;

pub use crate::style_helpers::TaffyAuto;

pub use crate::style_helpers::TaffyFitContent;

pub use crate::style_helpers::TaffyMaxContent;

pub use crate::style_helpers::TaffyMinContent;

pub use crate::style_helpers::TaffyZero;

pub use crate::tree::LayoutTree;

Enums

AlignContent

Sets the distribution of space between and around content items For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

AlignItems

Used to control how child Nodes are aligned. For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

AvailableSpace

The amount of space available to a node in a given axis https://www.w3.org/TR/css-sizing-3/#available

Dimension

A unit of linear measurement

FlexDirection

The direction of the flexbox layout main axis.

FlexWrap

Controls whether flex items are forced onto one line or can wrap onto multiple lines.

LengthPercentage

A unit of linear measurement

LengthPercentageAuto

A unit of linear measurement

Functions

layout_flexbox

Apply the flexbox algorithm and recursively layout the specified node

Type Definitions

AlignSelf

Used to control how the specified Nodes is aligned. Overrides the parent Node’s AlignItems property. For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

JustifyContent

Sets the distribution of space between and around content items For Flexbox it controls alignment in the main axis For Grid it controls alignment in the inline axis

JustifyItems

Used to control how child Nodes are aligned. Does not apply to Flexbox, and will be ignored if specified on a flex container For Grid it controls alignment in the inline axis

JustifySelf

Used to control how the specified Nodes is aligned. Overrides the parent Node’s JustifyItems property. Does not apply to Flexbox, and will be ignored if specified on a flex child For Grid it controls alignment in the inline axis