taffy

Struct Style

pub struct Style<S: CheapCloneStr = String> {

    pub dummy: PhantomData<S>,
    pub display: Display,
    pub item_is_table: bool,
    pub item_is_replaced: bool,
    pub box_sizing: BoxSizing,
    pub direction: Direction,
    pub overflow: Point<Overflow>,
    pub scrollbar_width: f32,
    pub contain: Contain,
    pub position: Position,
    pub inset: Rect<LengthPercentageAuto>,
    pub size: Size<Dimension>,
    pub min_size: Size<LengthPercentageAuto>,
    pub max_size: Size<LengthPercentageAuto>,
    pub aspect_ratio: Option<f32>,
    pub margin: Rect<LengthPercentageAuto>,
    pub padding: Rect<LengthPercentage>,
    pub border: Rect<LengthPercentage>,
    pub align_items: Option<AlignItems>,
    pub align_self: Option<AlignSelf>,
    pub justify_items: Option<AlignItems>,
    pub justify_self: Option<AlignSelf>,
    pub align_content: Option<AlignContent>,
    pub justify_content: Option<JustifyContent>,
    pub gap: Size<LengthPercentage>,
    pub grid_template_rows: Vec<GridTemplateComponent<S>>,
    pub grid_template_columns: Vec<GridTemplateComponent<S>>,
    pub grid_auto_rows: Vec<TrackSizingFunction>,
    pub grid_auto_columns: Vec<TrackSizingFunction>,
    pub grid_auto_flow: GridAutoFlow,
    pub grid_template_areas: Option<GridTemplateAreas<S>>,
    pub grid_template_column_names: Vec<Vec<S>>,
    pub grid_template_row_names: Vec<Vec<S>>,
    pub grid_row: Line<GridPlacement<S>>,
    pub grid_column: Line<GridPlacement<S>>,

}

A typed representation of the CSS style information for a single node.

The most important idea in flexbox is the notion of a “main” and “cross” axis, which are always perpendicular to each other. The orientation of these axes are controlled via the [FlexDirection] field of this struct.

This struct follows the CSS equivalent directly; information about the behavior on the web should transfer directly.

Detailed information about the exact behavior of each of these fields can be found on MDN by searching for the field name. The distinction between margin, padding and border is explained well in this introduction to the box model.

If the behavior does not match the flexbox layout algorithm on the web, please file a bug!

Fields
dummy: PhantomData<S>

This is a dummy field which is necessary to make Taffy compile with the grid feature disabled It should always be set to core::marker::PhantomData.

display: Display

What layout strategy should be used?

item_is_table: bool

Whether a child is display:table or not. This affects children of block layouts. This should really be part of Display, but it is currently seperate because table layout isn’t implemented

item_is_replaced: bool

Is it a replaced element like an image or form field? https://drafts.csswg.org/css-sizing-3/#min-content-zero

box_sizing: BoxSizing

Should size styles apply to the content box or the border box of the node

direction: Direction

Sets the direction of text, table and grid columns, and horizontal overflow.

overflow: Point<Overflow>

How children overflowing their container should affect layout

scrollbar_width: f32

How much space (in points) should be reserved for the scrollbars of Overflow::Scroll and Overflow::Auto nodes.

contain: Contain

The layout-affecting parts of the CSS contain property

position: Position

What should the position value of this struct use as a base offset?

inset: Rect<LengthPercentageAuto>

How should the position of this element be tweaked relative to the layout defined?

size: Size<Dimension>

Sets the initial size of the item

min_size: Size<LengthPercentageAuto>

Controls the minimum size of the item

max_size: Size<LengthPercentageAuto>

Controls the maximum size of the item

aspect_ratio: Option<f32>

Sets the preferred aspect ratio for the item

The ratio is calculated as width divided by height.

margin: Rect<LengthPercentageAuto>

How large should the margin be on each side?

padding: Rect<LengthPercentage>

How large should the padding be on each side?

border: Rect<LengthPercentage>

How large should the border be on each side?

align_items: Option<AlignItems>

How this node’s children aligned in the cross/block axis?

align_self: Option<AlignSelf>

How this node should be aligned in the cross/block axis Falls back to the parents AlignItems if not set

justify_items: Option<AlignItems>

How this node’s children should be aligned in the inline axis

justify_self: Option<AlignSelf>

How this node should be aligned in the inline axis Falls back to the parents JustifyItems if not set

align_content: Option<AlignContent>

How should content contained within this item be aligned in the cross/block axis

justify_content: Option<JustifyContent>

How should content contained within this item be aligned in the main/inline axis

gap: Size<LengthPercentage>

How large should the gaps between items in a grid or flex container be?

grid_template_rows: Vec<GridTemplateComponent<S>>

Defines the track sizing functions (heights) of the grid rows

grid_template_columns: Vec<GridTemplateComponent<S>>

Defines the track sizing functions (widths) of the grid columns

grid_auto_rows: Vec<TrackSizingFunction>

Defines the size of implicitly created rows

grid_auto_columns: Vec<TrackSizingFunction>

Defined the size of implicitly created columns

grid_auto_flow: GridAutoFlow

Controls how items get placed into the grid for auto-placed items

grid_template_areas: Option<GridTemplateAreas<S>>

Defines the rectangular grid areas

grid_template_column_names: Vec<Vec<S>>

The named lines between the columns

grid_template_row_names: Vec<Vec<S>>

The named lines between the rows

grid_row: Line<GridPlacement<S>>

Defines which row in the grid the item should start and end at

grid_column: Line<GridPlacement<S>>

Defines which column in the grid the item should start and end at

Implementations

impl<S: CheapCloneStr> Style<S>

pub const DEFAULT: Style<S>

The Default layout, in a form that can be used in const functions