taffy::style::alignment

Type Alias JustifySelf

pub type JustifySelf = AlignItems;

Controls alignment of an individual node.

Overrides the parent Node’s JustifyItems property. Does not apply to Flexbox, and will be ignored if specified on a flex child. For Grid it controls alignment in the inline axis.

MDN

Aliased Type

pub struct JustifySelf {
    pub keyword: AlignItemsKeyword,
    pub safety: AlignmentSafety,
}

Fields
keyword: AlignItemsKeyword

Position keyword.

safety: AlignmentSafety

Overflow-position modifier (safe / unsafe).