taffy::style

Module compact_length

A tagged-pointer abstraction that allows size styles in Taffy to be represented in just 64 bits. Wrapped by types in the super::dimension and super::grid modules.

Modules

compat 🔒

Note: these two functions are copied directly from the std (core) library. But by duplicating them here we can reduce MSRV from 1.84 all the way down to 1.65 while retaining const constructors and strict pointer provenance

inner 🔒

CompactLengthInner implementation for 64 bit platforms

Structs

CompactLength

A representation of a length as a compact 64-bit tagged pointer