taffy::style

Enum MinTrackSizingFunction

pub enum MinTrackSizingFunction {
    Fixed(LengthPercentage),
    MinContent,
    MaxContent,
    Auto,
}

Minimum track sizing function

Specifies the minimum size of a grid track. A grid track will automatically size between itโs minimum and maximum size based on the size of itโs contents, the amount of available space, and the sizing constraint the grid is being size under. See https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-columns

Variants

Fixed(LengthPercentage)

Track minimum size should be a fixed length or percentage value

MinContent

Track minimum size should be content sized under a min-content constraint

MaxContent

Track minimum size should be content sized under a max-content constraint

Auto

Track minimum size should be automatically sized

Implementations

impl MinTrackSizingFunction

pub fn is_intrinsic(&self) -> bool

Returns true if the min track sizing function is MinContent, MaxContent or Auto, else false.

pub fn definite_value(self, parent_size: Option<f32>) -> Option<f32>

Returns fixed point values directly. Attempts to resolve percentage values against the passed available_space and returns if this results in a concrete value (which it will if the available_space is Some). Otherwise returns None.

pub fn resolved_percentage_size(self, parent_size: f32) -> Option<f32>

Resolve percentage values against the passed parent_size, returning Some(value) Non-percentage values always return None.

pub fn uses_percentage(self) -> bool

Whether the track sizing functions depends on the size of the parent node