taffy::style

Module grid

Style types for CSS Grid layout

Structs

GridTemplateArea

Defines a grid area

GridTemplateAreas

Defines the value of the grid-template-areas property: the named areas plus the overall size (in tracks) of the area template.

GridTemplateRepetition

A typed representation of a repeat(..) in grid-template-* value

InvalidStringRepetitionValue

Error returned when trying to convert a string to a GridTrackRepetition and that string is not either “auto-fit” or “auto-fill”

MaxTrackSizingFunction

Maximum track sizing function

MinTrackSizingFunction

Minimum track sizing function

NamedGridLine

Defines a named grid line

Enums

ExpandedMaxTrackSizingFunction

The expanded, non-compact representation of a MaxTrackSizingFunction.

ExpandedMinTrackSizingFunction

The expanded, non-compact representation of a MinTrackSizingFunction.

GenericGridPlacement

A grid line placement specification which is generic over the coordinate system that it uses to define grid line positions.

GenericGridTemplateComponent

A type representing a component in a grid-template-* defintion where the type representing repeat()s is generic

GridAreaAxis 🔒

Axis as Row or Column

GridAreaEnd 🔒

Logical end (Start or End)

GridAutoFlow

Controls whether grid items are placed row-wise or column-wise. And whether the sparse or dense packing algorithm is used.

GridPlacement

A grid line placement specification. Used for grid-[row/column]-[start/end]. Named tracks are not implemented.

GridTemplateComponent

An element in a grid-template-columns or grid-template-rows definition. Either a track sizing function or a repeat().

RepetitionCount

The first argument to a repeated track definition. This type represents the type of automatic repetition to perform.

Traits

GenericRepetition

A trait to represent a repeat() clause in a grid-template-* definition

GridContainerStyle

The set of styles required for a CSS Grid container

GridItemStyle

The set of styles required for a CSS Grid item (child of a CSS Grid container)

TemplateLineNames

A nested list of line names. This is effectively a generic representation of Vec<Vec<String>> that allows both the collection and string type to be customised.

Type Aliases

NonNamedGridPlacement 🔒

A grid line placement using CSS grid line coordinates to specify line positions. This uses the same coordinate system as the public GridPlacement type but doesn’t support named lines (these are expected to have already been resolved by the time values of this type are constructed).

OriginZeroGridPlacement 🔒

A grid line placement using the normalized OriginZero coordinates to specify line positions.

TrackSizingFunction

The sizing function for a grid track (row/column)