taffy::style::grid

Type Alias NonNamedGridPlacement

pub(crate) type NonNamedGridPlacement = GenericGridPlacement<GridLine>;

A grid line placement using CSS grid line coordinates to specify line positions. This uses the same coordinate system as the public GridPlacement type but doesn’t support named lines (these are expected to have already been resolved by the time values of this type are constructed).

Aliased Type

pub(crate) enum NonNamedGridPlacement {
    Auto,
    Line(GridLine),
    Span(u16),
}

Variants

Auto

Place item according to the auto-placement algorithm, and the parent’s grid_auto_flow property

Line(GridLine)

Place item at specified line (column or row) index

Span(u16)

Item should span specified number of tracks (columns or rows)