taffy::style

Struct CompactLength

#[repr(transparent)]
pub struct CompactLength(CompactLengthInner);

A representation of a length as a compact 64-bit tagged pointer

Tuple Fields
0: CompactLengthInner
Implementations

impl CompactLength

pub const CALC_TAG: usize = 0b000

The tag indicating a calc() value

pub const LENGTH_TAG: usize = 0b0000_0001

The tag indicating a length value

pub const PERCENT_TAG: usize = 0b0000_0010

The tag indicating a percentage value

pub const AUTO_TAG: usize = 0b0000_0011

The tag indicating an auto value

pub const FR_TAG: usize = 0b0000_0100

The tag indicating an fr value

pub const MIN_CONTENT_TAG: usize = 0b00000111

The tag indicating a min-content value

pub const MAX_CONTENT_TAG: usize = 0b00001111

The tag indicating a max-content value

pub const FIT_CONTENT_PX_TAG: usize = 0b00010111

The tag indicating a fit-content value with px limit

pub const FIT_CONTENT_PERCENT_TAG: usize = 0b00011111

The tag indicating a fit-content value with percent limit

pub const FIT_CONTENT_KEYWORD_TAG: usize = 0b00100111

The tag indicating a plain fit-content keyword value (no limit)

pub const STRETCH_TAG: usize = 0b00101111

The tag indicating a stretch keyword value

pub const CONTENT_TAG: usize = 0b00110111

The tag indicating a content keyword value

impl CompactLength

pub const fn length(val: f32) -> Self

An absolute length in some abstract units. Users of Taffy may define what they correspond to in their application (pixels, logical pixels, mm, etc) as they see fit.

pub const fn percent(val: f32) -> Self

A percentage length relative to the size of the containing block.

NOTE: percentages are represented as a f32 value in the range [0.0, 1.0] NOT the range [0.0, 100.0]

pub fn calc(ptr: *const ()) -> Self

A calc() value. The value passed here is treated as an opaque handle to the actual calc representation and may be a pointer, index, etc.

The low 3 bits are used as a tag value and will be returned as 0.

pub const fn auto() -> Self

The dimension should be automatically computed according to algorithm-specific rules regarding the default size of boxes.

pub const fn fr(val: f32) -> Self

The dimension as a fraction of the total available grid space (fr units in CSS) Specified value is the numerator of the fraction. Denominator is the sum of all fraction specified in that grid dimension Spec: https://www.w3.org/TR/css3-grid-layout/#fr-unit

pub const fn min_content() -> Self

The size should be the “min-content” size. This is the smallest size that can fit the item’s contents with ALL soft line-wrapping opportunities taken

pub const fn max_content() -> Self

The size should be the “max-content” size. This is the smallest size that can fit the item’s contents with NO soft line-wrapping opportunities taken

pub const fn fit_content_px(limit: f32) -> Self

The size should be computed according to the “fit content” formula: max(min_content, min(max_content, limit)) where:

min_content is the min-content size

max_content is the max-content size

limit is a LENGTH value passed to this function

The effect of this is that the item takes the size of limit clamped by the min-content and max-content sizes.

pub const fn fit_content_percent(limit: f32) -> Self

The size should be computed according to the “fit content” formula: max(min_content, min(max_content, limit)) where:

min_content is the min-content size

max_content is the max-content size

limit is a PERCENTAGE value passed to this function

The effect of this is that the item takes the size of limit clamped by the min-content and max-content sizes.

pub const fn fit_content_keyword() -> Self

The size should be computed according to the “fit content” formula: max(min_content, min(max_content, stretch)) where:

min_content is the min-content size

max_content is the max-content size

stretch is the “stretch-fit” size (the size the box would take if it filled the available space)

pub const fn stretch() -> Self

The size should be the “stretch-fit” size: the size the box would take if it filled the available space (https://www.w3.org/TR/css-sizing-4/#stretch-fit-sizing)

pub const fn content() -> Self

The size should be an automatic size based on the box’s content (https://www.w3.org/TR/css-flexbox-1/#valdef-flex-basis-content)

This keyword is only valid for flex-basis. In any other context it behaves as auto.

pub fn tag(self) -> usize

Get the primary tag

pub fn value(self) -> f32

Get the numeric value associated with the CompactLength (e.g. the pixel value for a LENGTH variant)

pub fn calc_value(self) -> *const ()

Get the calc pointer of the CompactLength

pub fn is_calc(self) -> bool

Returns true if the value is 0 px

pub fn is_zero(self) -> bool

Returns true if the value is 0 px

pub fn is_length_or_percentage(self) -> bool

Returns true if the value is a length or percentage value

pub fn is_auto(self) -> bool

Returns true if the value is auto

pub fn is_content(self) -> bool

Returns true if the value is the content keyword

pub fn is_min_content(self) -> bool

Returns true if the value is min-content

pub fn is_max_content(self) -> bool

Returns true if the value is max-content

pub fn is_fit_content(self) -> bool

Returns true if the value is a fit-content(…) value

pub fn is_sizing_keyword(self) -> bool

Returns true if the value is min-content, max-content, fit-content, fit-content(…), or stretch

pub fn is_max_or_fit_content(self) -> bool

Returns true if the value is max-content or a fit-content(…) value

pub fn is_max_content_alike(&self) -> bool

Returns true if the max track sizing function is MaxContent, FitContent or Auto else false. “In all cases, treat auto and fit-content() as max-content, except where specified otherwise for fit-content().” See: https://www.w3.org/TR/css-grid-1/#algo-terms

pub fn is_min_or_max_content(&self) -> bool

Returns true if the min track sizing function is MinContent or MaxContent, else false.

pub fn is_intrinsic(self) -> bool

Returns true if the value is auto, min-content, max-content, or fit-content(…)

pub fn is_fr(self) -> bool

Returns true if the value is and fr value

pub fn uses_percentage(self) -> bool

Whether the track sizing functions depends on the size of the parent node

pub fn resolved_percentage_size( self, parent_size: f32, calc_resolver: impl Fn(*const (), f32) -> f32, ) -> Option<f32>

Resolve percentage values against the passed parent_size, returning Some(value) Non-percentage values always return None.