taffy::style

Struct LengthPercentage

pub struct LengthPercentage(pub(crate) CompactLength);

A unit of linear measurement

This is commonly combined with Rect, Point and Size<T>.

Tuple Fields
0: CompactLength
Implementations

impl LengthPercentage

pub const fn length(val: f32) -> Self

An absolute length in some abstract units. Users of Taffy may define what they correspond to in their application (pixels, logical pixels, mm, etc) as they see fit.

pub const fn percent(val: f32) -> Self

A percentage length relative to the size of the containing block.

NOTE: percentages are represented as a f32 value in the range [0.0, 1.0] NOT the range [0.0, 100.0]

pub fn calc(ptr: *const ()) -> Self

A calc() value. The value passed here is treated as an opaque handle to the actual calc representation and may be a pointer, index, etc.

The low 3 bits are used as a tag value and will be returned as 0.

pub const unsafe fn from_raw(val: CompactLength) -> Self

Create a LengthPercentage from a raw CompactLength.

Safety

CompactLength must represent a valid variant for LengthPercentage

pub const fn into_raw(self) -> CompactLength

Get the underlying CompactLength representation of the value

pub fn expand(self) -> ExpandedLengthPercentage

Expand the compact representation into an ExpandedLengthPercentage enum.

This is useful when integrating with other libraries (e.g. for style inspection or serialization) as it allows the value to be pattern-matched without having to work with the raw CompactLength tagged-pointer representation directly.