taffy::style

Struct LengthPercentageAuto

pub struct LengthPercentageAuto(/* private fields */);

A unit of linear measurement

This is commonly combined with Rect, Point and Size<T>.

Implementations

impl LengthPercentageAuto

pub const fn length(val: f32) -> Self

An absolute length in some abstract units. Users of Taffy may define what they correspond to in their application (pixels, logical pixels, mm, etc) as they see fit.

pub const fn percent(val: f32) -> Self

A percentage length relative to the size of the containing block.

NOTE: percentages are represented as a f32 value in the range [0.0, 1.0] NOT the range [0.0, 100.0]

pub const fn auto() -> Self

The dimension should be automatically computed according to algorithm-specific rules regarding the default size of boxes.

pub fn calc(ptr: *const ()) -> Self

A calc() value. The value passed here is treated as an opaque handle to the actual calc representation and may be a pointer, index, etc.

The low 3 bits are used as a tag value and will be returned as 0.

pub unsafe fn from_raw(val: CompactLength) -> Self

Create a LengthPercentageAuto from a raw CompactLength.

Safety

CompactLength must represent a valid variant for LengthPercentageAuto

pub fn into_raw(self) -> CompactLength

Get the underlying CompactLength representation of the value

pub fn resolve_to_option( self, context: f32, calc_resolver: impl Fn(*const (), f32) -> f32, ) -> Option<f32>

Returns:

Some(length) for Length variants

Some(resolved) using the provided context for Percent variants

None for Auto variants

pub fn is_auto(self) -> bool

Returns true if value is LengthPercentageAuto::Auto