taffy::style

Struct MinTrackSizingFunction

pub struct MinTrackSizingFunction(pub(crate) CompactLength);

Minimum track sizing function

Specifies the minimum size of a grid track. A grid track will automatically size between it’s minimum and maximum size based on the size of it’s contents, the amount of available space, and the sizing constraint the grid is being size under. See https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-columns

Tuple Fields
0: CompactLength
Implementations

impl MinTrackSizingFunction

pub const fn length(val: f32) -> Self

An absolute length in some abstract units. Users of Taffy may define what they correspond to in their application (pixels, logical pixels, mm, etc) as they see fit.

pub const fn percent(val: f32) -> Self

A percentage length relative to the size of the containing block.

NOTE: percentages are represented as a f32 value in the range [0.0, 1.0] NOT the range [0.0, 100.0]

pub const fn auto() -> Self

The dimension should be automatically computed according to algorithm-specific rules regarding the default size of boxes.

pub const fn min_content() -> Self

The size should be the “min-content” size. This is the smallest size that can fit the item’s contents with ALL soft line-wrapping opportunities taken

pub const fn max_content() -> Self

The size should be the “max-content” size. This is the smallest size that can fit the item’s contents with NO soft line-wrapping opportunities taken

pub fn calc(ptr: *const ()) -> Self

A calc() value. The value passed here is treated as an opaque handle to the actual calc representation and may be a pointer, index, etc.

The low 3 bits are used as a tag value and will be returned as 0.

pub unsafe fn from_raw(val: CompactLength) -> Self

Create a LengthPercentageAuto from a raw CompactLength.

Safety

CompactLength must represent a valid variant for LengthPercentageAuto

pub fn into_raw(self) -> CompactLength

Get the underlying CompactLength representation of the value

pub fn is_intrinsic(&self) -> bool

Returns true if the min track sizing function is MinContent, MaxContent or Auto, else false.

pub fn is_min_or_max_content(&self) -> bool

Returns true if the min track sizing function is MinContent or MaxContent, else false.

pub fn is_fr(&self) -> bool

Returns true if the value is an fr value

pub fn is_auto(&self) -> bool

Returns true if the is Auto, else false.

pub fn is_min_content(&self) -> bool

Returns true if value is MinContent

pub fn is_max_content(&self) -> bool

Returns true if value is MaxContent

pub fn definite_value( self, parent_size: Option<f32>, calc_resolver: impl Fn(*const (), f32) -> f32, ) -> Option<f32>

Returns fixed point values directly. Attempts to resolve percentage values against the passed available_space and returns if this results in a concrete value (which it will if the available_space is Some). Otherwise returns None.

pub fn resolved_percentage_size( self, parent_size: f32, calc_resolver: impl Fn(*const (), f32) -> f32, ) -> Option<f32>

Resolve percentage values against the passed parent_size, returning Some(value) Non-percentage values always return None.

pub fn uses_percentage(self) -> bool

Whether the track sizing functions depends on the size of the parent node

pub fn expand(self) -> ExpandedMinTrackSizingFunction

Expand the compact representation into an ExpandedMinTrackSizingFunction enum.

This is useful when integrating with other libraries (e.g. for style inspection or serialization) as it allows the value to be pattern-matched without having to work with the raw CompactLength tagged-pointer representation directly.