taffy::style

Trait FlexboxContainerStyle

pub trait FlexboxContainerStyle: CoreStyle {
    // Provided methods
    fn flex_direction(&self) -> FlexDirection { ... }
    fn flex_wrap(&self) -> FlexWrap { ... }
    fn gap(&self) -> Size<LengthPercentage> { ... }
    fn align_content(&self) -> Option<AlignContent> { ... }
    fn align_items(&self) -> Option<AlignItems> { ... }
    fn justify_content(&self) -> Option<JustifyContent> { ... }
}

The set of styles required for a Flexbox container

Provided Methods

fn flex_direction(&self) -> FlexDirection

Which direction does the main axis flow in?

fn flex_wrap(&self) -> FlexWrap

Should elements wrap, or stay in a single line?

fn gap(&self) -> Size<LengthPercentage>

How large should the gaps between items in a grid or flex container be?

fn align_content(&self) -> Option<AlignContent>

How should content contained within this item be aligned in the cross/block axis

fn align_items(&self) -> Option<AlignItems>

How this node’s children aligned in the cross/block axis?

fn justify_content(&self) -> Option<JustifyContent>

How this node’s children should be aligned in the inline axis