taffy::style

Trait FlexboxItemStyle

pub trait FlexboxItemStyle: CoreStyle {
    // Provided methods
    fn flex_basis(&self) -> Dimension { ... }
    fn flex_grow(&self) -> f32 { ... }
    fn flex_shrink(&self) -> f32 { ... }
    fn align_self(&self) -> Option<AlignSelf> { ... }
}

The set of styles required for a Flexbox item (child of a Flexbox container)

Provided Methods

fn flex_basis(&self) -> Dimension

Sets the initial main axis size of the item

fn flex_grow(&self) -> f32

The relative rate at which this item grows when it is expanding to fill space

fn flex_shrink(&self) -> f32

The relative rate at which this item shrinks when it is contracting to fit into space

fn align_self(&self) -> Option<AlignSelf>

How this node should be aligned in the cross/block axis Falls back to the parents AlignItems if not set