Type Alias taffy::style::NonRepeatedTrackSizingFunction

pub type NonRepeatedTrackSizingFunction = MinMax<MinTrackSizingFunction, MaxTrackSizingFunction>;

The sizing function for a grid track (row/column) (either auto-track or template track) May either be a MinMax variant which specifies separate values for the min-/max- track sizing functions or a scalar value which applies to both track sizing functions.

Aliased Type

struct NonRepeatedTrackSizingFunction {
    pub min: MinTrackSizingFunction,
    pub max: MaxTrackSizingFunction,
}

Fields
min: MinTrackSizingFunction

The value representing the minimum

max: MaxTrackSizingFunction

The value representing the maximum

Implementations

impl NonRepeatedTrackSizingFunction

pub fn min_sizing_function(&self) -> MinTrackSizingFunction

Extract the min track sizing function

pub fn max_sizing_function(&self) -> MaxTrackSizingFunction

Extract the max track sizing function

pub fn has_fixed_component(&self) -> bool

Determine whether at least one of the components (“min” and “max”) are fixed sizing function