Module taffy::style_helpers

Helper functions which it make it easier to create instances of types in the style and geometry modules.

Traits

FromFlex

Trait to create constant percent values from plain numbers

FromPercent

Trait to create constant percent values from plain numbers

FromPoints

Trait to create constant points values from plain numbers

TaffyAuto

Trait to abstract over auto values

TaffyFitContent

Trait to create constant points values from plain numbers

TaffyGridLine

Trait to abstract over grid line values

TaffyGridSpan

Trait to abstract over grid span values

TaffyMaxContent

Trait to abstract over max_content values

TaffyMinContent

Trait to abstract over min_content values

TaffyZero

Trait to abstract over zero values

Functions

auto

Returns the auto value for that type

fit_content

Returns a value of the inferred type which represent a constant of points

line

Specifies a grid line to place a grid item between in CSS Grid Line coordinates:

max_content

Returns the auto value for that type

min_content

Returns the auto value for that type

percent

Returns a value of the inferred type which represent a constant of points

points

Returns a value of the inferred type which represent a constant of points

span

Returns a GridPlacement::Span

zero

Returns the zero value for that type