taffy::tree::traits

Trait CacheTree

pub trait CacheTree {
    // Required methods
    fn cache_get(
        &mut self,
        node_id: NodeId,
        input: &LayoutInput,
    ) -> Option<LayoutOutput>;
    fn cache_store(
        &mut self,
        node_id: NodeId,
        input: &LayoutInput,
        layout_output: LayoutOutput,
    );
    fn cache_clear(&mut self, node_id: NodeId);
}

Trait used by the compute_cached_layout method which allows cached layout results to be stored and retrieved.

The Cache struct implements a per-node cache that is compatible with this trait.

Required Methods

fn cache_get( &mut self, node_id: NodeId, input: &LayoutInput, ) -> Option<LayoutOutput>

Try to retrieve a cached result from the cache

fn cache_store( &mut self, node_id: NodeId, input: &LayoutInput, layout_output: LayoutOutput, )

Store a computed size in the cache

fn cache_clear(&mut self, node_id: NodeId)

Clear all cache entries for the node