taffy

Module util

Helpful misc. utilities such as a function to debug print a tree

Modules

math 🔒

Contains numerical helper traits and functions

print 🔒

Contains functions for printing a debug representation of the tree

resolve 🔒

Helper trait to calculate dimensions during layout resolution

sys 🔒

Allocator-flexible data types

Traits

MaybeMath

A trait to conveniently calculate minimums and maximums when some data may not be defined

MaybeResolve

Trait to encapsulate behaviour where we need to resolve from a potentially context-dependent size or dimension into a context-independent size or dimension.

ResolveOrZero

Trait to encapsulate behaviour where we need to resolve from a potentially context-dependent size or dimension into a context-independent size or dimension.

Functions

print_tree

Prints a debug representation of the computed layout for a tree of nodes, starting with the passed root node.

write_tree

Writes a debug representation of the computed layout to the writer, starting with the passed root node.