taffy::util::sys

Module std

For when std is enabled

Macros

format 🔒

Creates a String using interpolation of runtime expressions.

Functions

abs 🔒

Computes the absolute value

ceil 🔒

Rounds up to the nearest whole number

f32_max 🔒

Returns the largest of two f32 values

f32_min 🔒

Returns the smallest of two f32 values

floor 🔒

Rounds down to the nearest whole number

new_vec_with_capacity 🔒

Creates a new vector with the capacity for the specified number of items before it must be resized

round 🔒

Rounds to the nearest whole number

Type Aliases

ChildrenVec 🔒

A vector of child nodes

DefaultCheapStr 🔒

The default type for representing strings in Taffy styles

GridTrackVec 🔒

A vector of grid tracks

Map 🔒

A map

String 🔒

A string

Vec 🔒

An allocation-backend agnostic vector type