taffy::compute::grid

Module placement

Implements placing items in the grid and resolving the implicit grid. https://www.w3.org/TR/css-grid-1/#placement

Functions

place_definite_grid_item 🔒

8.5. Grid Item Placement Algorithm Place a single definitely placed item into the grid

place_definite_secondary_axis_item 🔒

8.5. Grid Item Placement Algorithm Step 2. Place remaining children with definite secondary axis positions

place_grid_items 🔒

8.5. Grid Item Placement Algorithm Place items into the grid, generating new rows/column into the implicit grid as required

place_indefinitely_positioned_item 🔒

8.5. Grid Item Placement Algorithm Step 4. Position the remaining grid items.

record_grid_placement 🔒

Record the grid item in both CellOccupancyMatric and the GridItems list once a definite placement has been determined