taffy::compute::grid

Module types

Structs and enums that are used within the grid module

Modules

cell_occupancy 🔒

Contains CellOccupancyMatrix used to track occupied cells during grid placement

coordinates 🔒

Taffy uses two coordinate systems to refer to grid lines (the gaps/gutters between rows/columns):

grid_item 🔒

Contains GridItem used to represent a single grid item during layout

grid_track 🔒

Contains GridTrack used to represent a single grid track (row/column) during layout

grid_track_counts 🔒

Contains TrackCounts used to keep track of the number of tracks in the explicit and implicit grids. Also contains coordinate conversion functions which depend on those counts

named 🔒

Code for resolving name grid lines and areas