Module taffy::style

A representation of CSS layout properties in Rust, used for flexbox layout

Structs

Style

The flexbox layout information for a single Node.

Enums

AlignContent

Sets the distribution of space between and around content items For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

AlignItems

Used to control how child Nodes are aligned. For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

AvailableSpace

The amount of space available to a node in a given axis https://www.w3.org/TR/css-sizing-3/#available

Dimension

A unit of linear measurement

Display

Sets the layout used for the children of this node

FlexDirection

The direction of the flexbox layout main axis.

FlexWrap

Controls whether flex items are forced onto one line or can wrap onto multiple lines.

GridAutoFlow

Controls whether grid items are placed row-wise or column-wise. And whether the sparse or dense packing algorithm is used.

GridTrackRepetition

The first argument to a repeated track definition. This type represents the type of automatic repetition to perform.

LengthPercentage

A unit of linear measurement

LengthPercentageAuto

A unit of linear measurement

MaxTrackSizingFunction

Maximum track sizing function

MinTrackSizingFunction

Minimum track sizing function

Position

The positioning strategy for this item.

TrackSizingFunction

The sizing function for a grid track (row/column) See https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-columns

Type Aliases

AlignSelf

Used to control how the specified Nodes is aligned. Overrides the parent Node’s AlignItems property. For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

GridPlacement

A grid line placement specification. Used for grid-[row/column]-[start/end]. Named tracks are not implemented.

JustifyContent

Sets the distribution of space between and around content items For Flexbox it controls alignment in the main axis For Grid it controls alignment in the inline axis

JustifyItems

Used to control how child Nodes are aligned. Does not apply to Flexbox, and will be ignored if specified on a flex container For Grid it controls alignment in the inline axis

JustifySelf

Used to control how the specified Nodes is aligned. Overrides the parent Node’s JustifyItems property. Does not apply to Flexbox, and will be ignored if specified on a flex child For Grid it controls alignment in the inline axis

NonRepeatedTrackSizingFunction

The sizing function for a grid track (row/column) (either auto-track or template track) May either be a MinMax variant which specifies separate values for the min-/max- track sizing functions or a scalar value which applies to both track sizing functions.