Module taffy::style

A representation of CSS layout properties in Rust, used for flexbox layout

Structs

Style

The flexbox layout information for a single Node.

Enums

AlignContent

Sets the distribution of space between and around content items along the cross-axis

AlignItems

How Nodes are aligned relative to the cross axis

AlignSelf

Overrides the inherited AlignItems behavior for this node.

Dimension

A unit of linear measurement

Display

Sets the layout used for the children of this node

FlexDirection

The direction of the flexbox layout main axis.

FlexWrap

Controls whether flex items are forced onto one line or can wrap onto multiple lines.

JustifyContent

Sets the distribution of space between and around content items along the main-axis

PositionType

The positioning strategy for this item.