Type Alias taffy::style::JustifyItems

pub type JustifyItems = AlignItems;

Used to control how child Nodes are aligned. Does not apply to Flexbox, and will be ignored if specified on a flex container For Grid it controls alignment in the inline axis

MDN

Aliased Type

enum JustifyItems {
    Start,
    End,
    FlexStart,
    FlexEnd,
    Center,
    Baseline,
    Stretch,
}

Variants

Start

Items are packed toward the start of the axis

End

Items are packed toward the end of the axis

FlexStart

Items are packed towards the flex-relative start of the axis.

For flex containers with flex_direction RowReverse or ColumnReverse this is equivalent to End. In all other cases it is equivalent to Start.

FlexEnd

Items are packed towards the flex-relative end of the axis.

For flex containers with flex_direction RowReverse or ColumnReverse this is equivalent to Start. In all other cases it is equivalent to End.

Center

Items are packed along the center of the cross axis

Baseline

Items are aligned such as their baselines align

Stretch

Stretch to fill the container