taffy

Module tree

Contains both a high-level interface to Taffy using a ready-made node tree, and a set of traits for defining custom node trees.

For documentation on the high-level API, see the [TaffyTree] struct.

For documentation on the low-level trait-based API, see the traits module.

Re-exports

pub use traits::LayoutPartialTree;

pub use traits::PrintTree;

pub use traits::RoundTree;

pub use traits::TraversePartialTree;

pub use traits::TraverseTree;

pub use traits::LayoutGridContainer;

Modules

cache 🔒

A cache for storing the results of layout computation

layout 🔒

Final data structures that represent the high-level UI layout

node 🔒

UI node types and related data structures.

traits

The abstractions that make up the core of Taffy’s low-level API

Structs

Cache

A cache for caching the results of a sizing a Grid Item or Flexbox Item

CollapsibleMarginSet

A set of margins that are available for collapsing with for block layout’s margin collapsing

Layout

The final result of a layout algorithm for a single node.

LayoutInput

A struct containing the inputs constraints/hints for laying out a node, which are passed in by the parent

LayoutOutput

A struct containing the result of laying a single node, which is returned up to the parent node

NodeId

A type representing the id of a single node in a tree of nodes

Enums

ClearState

Clear operation outcome. See Cache::clear

DetailedLayoutInfo

The additional information from layout algorithm

RequestedAxis

An axis that layout algorithms can be requested to compute a size for

RunMode

Whether we are performing a full layout, or we merely need to size the node

SizingMode

Whether styles should be taken into account when computing size