Struct taffy::tree::Layout

pub struct Layout {
    pub order: u32,
    pub location: Point<f32>,
    pub size: Size<f32>,
    pub content_size: Size<f32>,
    pub scrollbar_size: Size<f32>,
    pub border: Rect<f32>,
    pub padding: Rect<f32>,
}

The final result of a layout algorithm for a single node.

Fields
order: u32

The relative ordering of the node

Nodes with a higher order should be rendered on top of those with a lower order. This is effectively a topological sort of each tree.

location: Point<f32>

The top-left corner of the node

size: Size<f32>

The width and height of the node

content_size: Size<f32>

The width and height of the content inside the node. This may be larger than the size of the node in the case of overflowing content and is useful for computing a โscroll width/heightโ for scrollable nodes

scrollbar_size: Size<f32>

The size of the scrollbars in each dimension. If there is no scrollbar then the size will be zero.

border: Rect<f32>

The size of the borders of the node

padding: Rect<f32>

The size of the padding of the node

Implementations

impl Layout

pub const fn new() -> Self

Creates a new zero-Layout.

The Zero-layout has size and location set to ZERO. The order value of this layout is set to the minimum value of 0. This means it should be rendered below all other Layouts.

pub const fn with_order(order: u32) -> Self

Creates a new zero-Layout with the supplied order value.

Nodes with a higher order should be rendered on top of those with a lower order. The Zero-layout has size and location set to ZERO.

impl Layout

pub fn scroll_width(&self) -> f32

Return the scroll width of the node. The scroll width is the difference between the width and the content width, floored at zero

pub fn scroll_height(&self) -> f32

Return the scroll width of the node. The scroll width is the difference between the width and the content width, floored at zero