Struct taffy::tree::TaffyTree

pub struct TaffyTree<NodeContext = ()> { /* private fields */ }

An entire tree of UI nodes. The entry point to Taffy’s high-level API.

Allows you to build a tree of UI nodes, run Taffy’s layout algorithms over that tree, and then access the resultant layout.

Implementations

impl<NodeContext> TaffyTree<NodeContext>

pub fn new() -> Self

Creates a new TaffyTree

The default capacity of a TaffyTree is 16 nodes.

pub fn with_capacity(capacity: usize) -> Self

Creates a new TaffyTree that can store capacity nodes before reallocation

pub fn enable_rounding(&mut self)

Enable rounding of layout values. Rounding is enabled by default.

pub fn disable_rounding(&mut self)

Disable rounding of layout values. Rounding is enabled by default.

pub fn new_leaf(&mut self, layout: Style) -> TaffyResult<NodeId>

Creates and adds a new unattached leaf node to the tree, and returns the node of the new node

pub fn new_leaf_with_context( &mut self, layout: Style, context: NodeContext, ) -> TaffyResult<NodeId>

Creates and adds a new unattached leaf node to the tree, and returns the NodeId of the new node

Creates and adds a new leaf node with a supplied context

pub fn new_with_children( &mut self, layout: Style, children: &[NodeId], ) -> TaffyResult<NodeId>

Creates and adds a new node, which may have any number of children

pub fn clear(&mut self)

Drops all nodes in the tree

pub fn remove(&mut self, node: NodeId) -> TaffyResult<NodeId>

Remove a specific node from the tree and drop it

Returns the id of the node removed.

pub fn set_node_context( &mut self, node: NodeId, measure: Option<NodeContext>, ) -> TaffyResult<()>

Sets the context data associated with the node

pub fn get_node_context(&self, node: NodeId) -> Option<&NodeContext>

Gets a reference to the the context data associated with the node

pub fn get_node_context_mut(&mut self, node: NodeId) -> Option<&mut NodeContext>

Gets a mutable reference to the the context data associated with the node

pub fn add_child(&mut self, parent: NodeId, child: NodeId) -> TaffyResult<()>

Adds a child node under the supplied parent

pub fn insert_child_at_index( &mut self, parent: NodeId, child_index: usize, child: NodeId, ) -> TaffyResult<()>

Inserts a child node at the given child_index under the supplied parent, shifting all children after it to the right.

pub fn set_children( &mut self, parent: NodeId, children: &[NodeId], ) -> TaffyResult<()>

Directly sets the children of the supplied parent

pub fn remove_child( &mut self, parent: NodeId, child: NodeId, ) -> TaffyResult<NodeId>

Removes the child of the parent node

The child is not removed from the tree entirely, it is simply no longer attached to its previous parent.

pub fn remove_child_at_index( &mut self, parent: NodeId, child_index: usize, ) -> TaffyResult<NodeId>

Removes the child at the given index from the parent

The child is not removed from the tree entirely, it is simply no longer attached to its previous parent.

pub fn replace_child_at_index( &mut self, parent: NodeId, child_index: usize, new_child: NodeId, ) -> TaffyResult<NodeId>

Replaces the child at the given child_index from the parent node with the new child node

The child is not removed from the tree entirely, it is simply no longer attached to its previous parent.

pub fn child_at_index( &self, parent: NodeId, child_index: usize, ) -> TaffyResult<NodeId>

Returns the child node of the parent node at the provided child_index

pub fn total_node_count(&self) -> usize

Returns the total number of nodes in the tree

pub fn child_count(&self, parent: NodeId) -> TaffyResult<usize>

Returns the number of children of the parent node

pub fn children(&self, parent: NodeId) -> TaffyResult<Vec<NodeId>>

Returns a list of children that belong to the parent node

pub fn set_style(&mut self, node: NodeId, style: Style) -> TaffyResult<()>

Sets the Style of the provided node

pub fn style(&self, node: NodeId) -> TaffyResult<&Style>

Gets the Style of the provided node

pub fn layout(&self, node: NodeId) -> TaffyResult<&Layout>

Return this node layout relative to its parent

pub fn mark_dirty(&mut self, node: NodeId) -> TaffyResult<()>

Marks the layout computation of this node and its children as outdated

Performs a recursive depth-first search up the tree until the root node is reached

WARNING: this will stack-overflow if the tree contains a cycle

pub fn dirty(&self, node: NodeId) -> TaffyResult<bool>

Indicates whether the layout of this node (and its children) need to be recomputed

pub fn compute_layout_with_measure<MeasureFunction>( &mut self, node_id: NodeId, available_space: Size<AvailableSpace>, measure_function: MeasureFunction, ) -> Result<(), TaffyError>
where MeasureFunction: FnMut(Size<Option<f32>>, Size<AvailableSpace>, NodeId, Option<&mut NodeContext>) -> Size<f32>,

Updates the stored layout of the provided node and its children

pub fn compute_layout( &mut self, node: NodeId, available_space: Size<AvailableSpace>, ) -> Result<(), TaffyError>

Updates the stored layout of the provided node and its children

pub fn print_tree(&mut self, root: NodeId)

Prints a debug representation of the tree’s layout