taffy::util::sys

Module std

For when std is enabled

Macros

format ğŸ”’

Creates a String using interpolation of runtime expressions.

Functions

abs ğŸ”’

Computes the absolute value

ceil ğŸ”’

Rounds up to the nearest whole number

f32_max ğŸ”’

Returns the largest of two f32 values

f32_min ğŸ”’

Returns the smallest of two f32 values

floor ğŸ”’

Rounds down to the nearest whole number

new_vec_with_capacity ğŸ”’

Creates a new vector with the capacity for the specified number of items before it must be resized

round ğŸ”’

Rounds to the nearest whole number

single_value_vec ğŸ”’

Creates a new vector with the capacity for the specified number of items before it must be resized

Type Aliases

ChildrenVec ğŸ”’

A vector of child nodes

DefaultCheapStr ğŸ”’

The default type for representing strings in Taffy styles

GridTrackVec ğŸ”’

A vector of grid tracks

Map ğŸ”’

A map

String ğŸ”’

A string

Vec ğŸ”’

An allocation-backend agnostic vector type